Add files via upload
The project was done in partnership with an Australia based Data Analytics company Quantium. The data has been provided by them. It consists of potato chips sales across various stores in Australia from July 2018 - July 2019. Our task is to analyze sales of chips in this time period and recommend where improvements can be made. Python and it's data analysis libraries were applied to complete this task.

There are two datasets which have been provided to us, transactions and customer.